More Material You Styling Changes
This patch includes a WIP Redesign of the bottom action bar (This isn't the final design and the design will change over the course of the next few betas). The collapsing toolbar has been disabled in landscape mode to allow the player to be usable on phones. The Bottom navigation bar has also been shrunk in landscape to accommodate for smaller devices.